move cache outside of db
