+++ mushmom/database.py
@@ -2,8 +2,12 @@
 Functions related to database connection.  Currently using MongoDB,
 but could be replaced easily as long as functionality is the same
 
+Maintains a cache of db output to reduce calls to db when changes
+are unlikely.  Would cause issues if user makes changes in a different
+channel/server before submitting, but the user would basically know
+they were doing it.
+
 """
-import time
 import asyncio
 
 from motor.motor_asyncio import AsyncIOMotorClient
@@ -12,6 +16,7 @@ from datetime import datetime
 from typing import Optional, Union, Any, Hashable
 
 from . import config
+from .cache import TTLCache
 
 
 class Database:
@@ -46,7 +51,7 @@ class Database:
         self.guilds = self.db.guilds
 
         # caches
-        self.user_cache = DatabaseCache(seconds=60)  # 5 minute cache
+        self.user_cache = TTLCache(seconds=60)  # 5 minute cache
 
     async def get_user(
             self,
@@ -298,72 +303,3 @@ class Database:
 
     def close(self):  # not coroutine
         self.client.close()
-
-
-class DatabaseCache:
-    """
-    Maintains a cache of db output to reduce calls to db when changes
-    are unlikely.  Would cause issues if user makes changes in a different
-    channel/server before submitting, but the user would basically know
-    they were doing it.
-
-    Entries will expire after some time.  Verify auto calls every x calls
-    to get/add/remove
-
-    Parameters
-    ----------
-    seconds: int
-        the number of seconds to wait before expiring
-
-    """
-    def __init__(self, seconds: int):
-        super().__init__()
-        self.__ttl = seconds
-        self.__cache = {}
-
-        self.__cnt = 0
-        self.__recur = 20  # auto verify every 20 calls
-
-    def verify_cache_integrity(self) -> None:
-        """Loop through cache and remove all expired keys"""
-        current_time = time.monotonic()
-        to_remove = [k for (k, (v, t)) in self.__cache.items()
-                     if current_time > (t + self.__ttl)]
-        for k in to_remove:
-            del self.__cache[k]
-
-    def get(self, k: Hashable) -> Any:
-        self._verify_internal()
-
-        if k in self.__cache:
-            value, t = self.__cache.get(k)
-            current_time = time.monotonic()
-            if current_time <= (t + self.__ttl):
-                return value
-
-    def add(self, k: Hashable, value: Any) -> None:
-        self._verify_internal()
-        self.__cache[k] = (value, time.monotonic())
-
-    def remove(self, k: Hashable) -> None:
-        self._verify_internal()
-        self.__cache.pop(k, None)
-
-    def contains(self, k: Hashable) -> bool:
-        if k in self.__cache:
-            value, t = self.__cache.get(k)
-            current_time = time.monotonic()
-            return current_time <= (t + self.__ttl)
-        else:
-            return False
-
-    def __contains__(self, k: Hashable) -> bool:
-        return self.contains(k)
-
-    def _verify_internal(self):
-        """Verify every __recur calls"""
-        self.__cnt += 1
-
-        if self.__cnt >= self.__recur:
-            self.verify_cache_integrity()
-            self.__cnt = 0
